+++ src/test/java/org/ideplugins/vale_cli_plugin/actions/SettingsToolbarActionTest.java
@@ -1,19 +1,19 @@
-package org.ideplugins.vale_cli_plugin.action;
+package org.ideplugins.vale_cli_plugin.actions;
 
 import com.intellij.testFramework.fixtures.CodeInsightTestFixture;
 import org.ideplugins.vale_cli_plugin.BaseTest;
-import org.ideplugins.vale_cli_plugin.actions.ProjectSettingsToolbarAction;
 import org.ideplugins.vale_cli_plugin.testing.PluginTest;
 import org.ideplugins.vale_cli_plugin.testing.RunInEdtExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
+
 @PluginTest
 @ExtendWith(RunInEdtExtension.class)
-public class ProjectSettingsToolbarActionTest extends BaseTest{
+public class SettingsToolbarActionTest extends BaseTest{
 
     @Test
     public void testAction(CodeInsightTestFixture codeInsightTestFixture){
-        codeInsightTestFixture.testAction(new ProjectSettingsToolbarAction());
+        codeInsightTestFixture.testAction(new SettingsToolbarAction());
     }
 }